refactor: change structure of the project and change state data
