+++ src/components/ActiveQuiz/AnswerList/AnswerBlock.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { AnswerItem } from './AnswerItem/AnswerItem';
 import { AnswerDetails } from '../AnswerDetails/AnswerDetails';
+import './AnswerBlock.css';
 
 export const AnswerBlock = (props) => {
